+++ client/src/components/buttons/deleteProblemButton.tsx
@@ -9,20 +9,20 @@ interface Props {
 export const DeleteProblemButton = ({ id }: Props) => {
   const [deleteProblem] = useDeleteProblemMutation();
   const router = useRouter();
+
+  const handleDelete = async () => {
+    if (window.confirm('Are you sure? Deletion is permanent.')) {
+      await deleteProblem({
+        variables: { id },
+        update: (cache) => {
+          cache.evict({ id: 'Problem:' + id });
+        }
+      });
+      router.push('/problems');
+    }
+  };
   return (
-    <button
-      className='btn'
-      aria-label='Delete Problem'
-      onClick={async () => {
-        await deleteProblem({
-          variables: { id },
-          update: (cache) => {
-            cache.evict({ id: 'Problem:' + id });
-          }
-        });
-        router.push('/problems');
-      }}
-    >
+    <button className='btn' aria-label='Delete Problem' onClick={handleDelete}>
       <FaTrash size={22} />
     </button>
   );
